RBD: Set physical_block_size=4096
Ceph generally prefers a full 4K write where possible, without this on
FileStore the I/O submission thread stalls waiting for the kernel to
read the rest of the 4K object in from disk to make the entire 4K page
whole. On BlueStore, while it doesn't stall the submission thread it
will still have to read the partial object in due to the data checksum.

The infrastructure to configure this inside of nova/libvirt already
exists, however previously only the iSCSI/SolidFire driver set this
value. This extends the RBD driver to also handle this.

First during volume creation we set the provider_geometry data (which is
stored in the database).

Secondly during connection initialization we parse out this field into
the physical_block_size and logical_block_size information sent on the
API. The Nova driver consumes that information.